refactor: Move new/confirm assessment logic to shared location
These two apps will also be required by the youth risk assessment
which shares a lot of the framework logic.

This change moves the shared parts of these sub-apps to a more central
location so that the logic can be more easily shared.
