+++ app/person-escort-record/app/confirm/config.js
@@ -1,8 +1,6 @@
 module.exports = {
   hideSidebar: true,
-  name: 'create-person-escort-record',
-  journeyName: 'create-person-escort-record',
+  name: 'confirm-person-escort-record',
+  journeyName: 'confirm-person-escort-record',
   journeyPageTitle: 'person-escort-record::heading',
-  templatePath: 'person-escort-record/app/confirm/views/',
-  template: '../../../../form-wizard',
 }
