+++ app/person-escort-record/app/confirm/controllers.js
@@ -1,33 +1,7 @@
-const FormWizardController = require('../../../../common/controllers/form-wizard')
+const ConfirmAssessmentController = require('../../../../common/controllers/framework/confirm-assessment')
 const personEscortRecordService = require('../../../../common/services/person-escort-record')
 
-class ConfirmPersonEscortRecordController extends FormWizardController {
-  middlewareLocals() {
-    super.middlewareLocals()
-    this.use(this.setMoveId)
-  }
-
-  setMoveId(req, res, next) {
-    res.locals.moveId = req.move?.id
-    next()
-  }
-
-  middlewareChecks() {
-    super.middlewareChecks()
-    this.use(this.checkStatus)
-  }
-
-  checkStatus(req, res, next) {
-    const moveId = req.move?.id
-    const isCompleted = req?.assessment?.status === 'completed'
-
-    if (isCompleted) {
-      return next()
-    }
-
-    res.redirect(`/move/${moveId}`)
-  }
-
+class ConfirmPersonEscortRecordController extends ConfirmAssessmentController {
   async saveValues(req, res, next) {
     try {
       await personEscortRecordService.confirm(req.assessment.id)
@@ -35,13 +9,6 @@ class ConfirmPersonEscortRecordController extends FormWizardController {
     } catch (err) {
       next(err)
     }
-  }
-
-  successHandler(req, res) {
-    req.journeyModel.reset()
-    req.sessionModel.reset()
-
-    res.redirect(`/move/${req.move.id}`)
   }
 }
