+++ app/person-escort-record/app/new/config.js
@@ -1,7 +1,4 @@
 module.exports = {
-  hideSidebar: true,
   name: 'create-person-escort-record',
   journeyName: 'create-person-escort-record',
-  journeyPageTitle: 'person-escort-record::heading',
-  templatePath: 'person-escort-record/app/new/views/',
 }
