+++ app/person-escort-record/app/new/controllers.js
@@ -1,25 +1,9 @@
 const { get } = require('lodash')
 
-const FormWizardController = require('../../../../common/controllers/form-wizard')
+const NewAssessmentController = require('../../../../common/controllers/framework/new-assessment')
 const personEscortRecordService = require('../../../../common/services/person-escort-record')
 
-class NewPersonEscortRecordController extends FormWizardController {
-  middlewareChecks() {
-    super.middlewareChecks()
-    this.use(this.checkProfileExists)
-  }
-
-  checkProfileExists(req, res, next) {
-    if (req.move?.profile?.id) {
-      return next()
-    }
-
-    const error = new Error('Move profile not found')
-    error.statusCode = 404
-
-    next(error)
-  }
-
+class NewPersonEscortRecordController extends NewAssessmentController {
   async saveValues(req, res, next) {
     try {
       req.record = await personEscortRecordService.create(req.move.id)
@@ -33,13 +17,6 @@ class NewPersonEscortRecordController extends FormWizardController {
 
       next(err)
     }
-  }
-
-  successHandler(req, res) {
-    req.journeyModel.reset()
-    req.sessionModel.reset()
-
-    res.redirect(`/move/${req.move.id}`)
   }
 }
