+++ common/controllers/framework/index.js
@@ -1,9 +1,13 @@
+const ConfirmAssessmentController = require('./confirm-assessment')
 const frameworkOverviewController = require('./framework-overview')
 const FrameworkSectionController = require('./framework-section')
 const FrameworkStepController = require('./framework-step')
+const NewAssessmentController = require('./new-assessment')
 
 module.exports = {
+  ConfirmAssessmentController,
   frameworkOverviewController,
   FrameworkSectionController,
   FrameworkStepController,
+  NewAssessmentController,
 }
